+++ test/memhouse/ingest_relay_test.exs
@@ -168,6 +168,7 @@ defmodule MemHouse.IngestRelayTest do
 
     PromptCaptureProvider.start!([
       %{
+        "supporting_span" => "Melanie teaches watercolour on Tuesdays.",
         "statement" => "Melanie teaches watercolour on Tuesdays.",
         "kind" => "fact",
         "subject_type" => "peer",